api/requirements.txt
cryptography
flask
flask_cors
flask-migrate
flask-praetorian
flask-sqlalchemy
python-dotenv
pymysql
requests